+++ benchmark/bench/_template.js
