+++ benchmark/make-project/_template.js
@@ -1,0 +1,4 @@
+/**
+ * @param {URL} projectDir
+ */
+export async function run(projectDir) {}
